One-for-All-104 - I extracted AppFieldValue outside.
