+++ apps/web/src/feature/app/create/FieldResetButton.tsx
@@ -1,4 +1,4 @@
-import type { AppFieldValue } from "@feature/app/create/CreateAppClient";
+import type { AppFieldValue } from "@feature/app/create/app-field-value";
 
 export const FieldResetButton = ({
   setValue,
