+++ apps/web/src/feature/app/create/app-field-value.d.ts
@@ -1,0 +1,11 @@
+import type { FieldKind } from "@oneforall/domain/field/type";
+
+export type AppFieldValue = NonNullable<
+  Record<
+    number,
+    {
+      fieldName: string;
+      fieldKind: FieldKind;
+    }
+  >
+>;
